+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## master (to be 0.1.12)
 ### Enhancements
 - Speed up cache unzip by running them in parallel.
+- Remove the `still_download_sources` option. Instead, always download sources to avoid improper integration.
+- Fix resources integration (for ex. using `SwiftDate` as a static framework).
 
 ### Bug fixes
 None